+++ NEWS.md
@@ -23,6 +23,8 @@
   to use the new interface internally, so existing workflows should not break,
   but these are likely to be deprecated in the future.
 * Improved robustness against 0-length and gzipped input files
+* New helper function `read_long_sequence_table()` reads an OptimOTU-style long
+  sequence table from a tsv file.
 # optimotu.pipeline 0.6.2
 * Add option `force_denovo` in `pipeline_options.yaml` to force de novo
   clustering at some or all taxonomic ranks, i.e., to ignore taxonomic